% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/str_split_one.R
\name{str_split_one}
\alias{str_split_one}
\title{Split string}
\usage{
str_split_one(string, pattern, n = Inf)
}
\arguments{
\item{string}{A character vector with one element}

\item{pattern}{what to split on}
}
\value{
A character vector
}
\description{
Split string
}
\examples{
x <- "alpha,beta"
str_split_one(x, pattern=",")
str_split_one(x, pattern=",", n=2)
}
