% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/strsplit1.R
\name{strsplit1}
\alias{strsplit1}
\title{Split string}
\usage{
strsplit1(x, split)
}
\arguments{
\item{x}{A character vecor with one element}

\item{split}{what to split in}
}
\value{
A character vector
}
\description{
Split string
}
\examples{
x <- "alpha,beta"
strplit(x, ",")
}
